fix(striker-ui-api): break down access module
