+++ striker-ui-api/src/lib/accessModule/encrypt.ts
@@ -1,0 +1,10 @@
+import { sub } from './sub';
+
+export const encrypt: EncryptFunction = async (params) => {
+  const [result]: [Encrypted] = await sub('encrypt_password', {
+    params: [params],
+    pre: ['Account'],
+  });
+
+  return result;
+};
