+++ striker-ui-api/src/lib/accessModule/getAnvilData.ts
@@ -1,0 +1,8 @@
+import { getData } from './getData';
+import { sub } from './sub';
+
+export const getAnvilData = async () => {
+  await sub('get_anvils');
+
+  return getData<AnvilDataAnvilListHash>('anvils');
+};
