+++ striker-ui-api/src/lib/accessModule/getHostData.ts
@@ -1,0 +1,8 @@
+import { getData } from './getData';
+import { sub } from './sub';
+
+export const getHostData = async () => {
+  await sub('get_hosts');
+
+  return getData<AnvilDataHostListHash>('hosts');
+};
