+++ striker-ui-api/src/lib/accessModule/getUpsSpec.ts
@@ -1,0 +1,8 @@
+import { getData } from './getData';
+import { sub } from './sub';
+
+export const getUpsSpec = async () => {
+  await sub('get_ups_data', { pre: ['Striker'] });
+
+  return getData<AnvilDataUPSHash>('ups_data');
+};
