+++ striker-ui-api/src/lib/accessModule/variable.ts
@@ -1,0 +1,9 @@
+import { sub } from './sub';
+
+export const variable: InsertOrUpdateVariableFunction = async (params) => {
+  const [uuid]: [string] = await sub('insert_or_update_variables', {
+    params: [params],
+  });
+
+  return uuid;
+};
